[pci] Add a way to switch a PCI bus to PIO mode
In our current setup, Bus drivers are created before we know anything
about their roots, and roots reach back into the bus drivers to obtain
device configs while scanning downstream. Since the actual scan work
is done by the bus driver, but the information about the address space
of the tree is in the root, this method is being added. This should be
considered a hideous temporary hack that will go away as we move PCI to
userspace, but is needed in the short term to boostrap legacy PCI
enumeration for virtualized environments that use PIO based config space.

Change-Id: I6ce2c311ecb9c09361aeac8079b00e9354ff5f7a